Phase 5 infra: TinyStories prep, v0.2 config, review-gate fix
TinyStories provenance (datasets/SOURCES.md), scripts/prepare_tinystories.py,
configs/model_v0_2.yaml + train_tinystories.yaml (L6 h6 C192 block256, ~2.7M),
tests/test_prepare_tinystories{,_smoke}.py.

Correctly done: uint8 corpus storage (8x smaller than int64), a disk-space
precondition that fails loudly, and a Chinchilla-bounded TRAINING budget
(max_steps=3375 -> 55.3M tokens, matching the ~54.4M target for 2.72M params
per research/paper_notes/chinchilla.md).

get_batch (loader.py) now explicitly casts to int64 on read, so uint8-stored
data works without changing the existing Shakespeare (already-long) path.

Review-gate fix: SPEC.md's v0.2 summary row said vocab~65, contradicting its
own worked arithmetic three lines below (V=138, the real TinyStories char
count) -- corrected to ~138.

Known accepted trade-off (not blocking): prepare_tinystories.py still
downloads/tokenizes the full ~1.92GB corpus even though training only
samples ~55M of ~2B tokens (2.6%) -- one-time disk/bandwidth cost, not a
training-time or memory risk since max_steps already bounds the real run.

Tests: 45 passing.

+++ src/jimmylabs/data/loader.py
@@ -23,7 +23,7 @@ def get_batch(data: torch.Tensor, block_size: int, batch_size: int, device: str 
     x = torch.stack([data[i:i+block_size] for i in ix])
     y = torch.stack([data[i+1:i+block_size+1] for i in ix])
     
-    # Move to the desired device
-    x, y = x.to(device), y.to(device)
+    # Move to the desired device and ensure dtype is long (int64)
+    x, y = x.to(device, dtype=torch.long), y.to(device, dtype=torch.long)
     
     return x, y
